new : upgraded PHPUnit tests, dropped PHP <7.1 and Symfony 3.4
* fixed : removed composer.lock and modified .gitignore

It did not make sense to commit the `composer.lock` file.

* new : upgraded PHPUnit tests, dropped PHP <7.1 and Symfony 3.4

The tests of PHPUnit have been upgraded to more recent verions.
PHP 8 is now an accepted constraint. PHP versions prior to 7.1 are dropped.
Symfony 3.4 support is dropped as well.

* minor : inverted PHPUnit assertIsArray check

* fixed : test of SignalSpamStatistics dealing with message

SignalSpamStatisticsTest had to manager only SignalSpamStatistics. It was messing with Message as well.